+++ app/renderer/components/tabs/tabs.js
@@ -17,7 +17,6 @@ const windowActions = require('../../../../js/actions/windowActions')
 
 // State
 const windowState = require('../../../common/state/windowState')
-const tabUIState = require('../../../common/state/tabUIState')
 
 // Constants
 const dragTypes = require('../../../../js/constants/dragTypes')
@@ -129,7 +128,7 @@ class Tabs extends React.Component {
 
   mergeProps (state, ownProps) {
     const currentWindow = state.get('currentWindow')
-    const pageIndex = tabUIState.getPageIndex(currentWindow)
+    const pageIndex = frameStateUtil.getTabPageIndex(currentWindow)
     const tabsPerTabPage = Number(getSetting(settings.TABS_PER_PAGE))
     const startingFrameIndex = pageIndex * tabsPerTabPage
     const unpinnedTabs = frameStateUtil.getNonPinnedFrames(currentWindow) || Immutable.List()
